+++ src/features/mission-detail/components/mission-with-submission-history.tsx
@@ -7,8 +7,8 @@ import { SubmissionHistoryWrapper } from "@/features/mission-detail/components/s
 import type { SubmissionData } from "@/features/mission-detail/types/detail-types";
 import { MissionGuidanceArrow } from "@/features/missions/components/mission-guidance-arrow";
 import { useMissionSubmission } from "@/features/missions/hooks/use-mission-submission";
-import { ARTIFACT_TYPES } from "@/lib/artifactTypes";
 import { createClient } from "@/lib/supabase/client";
+import { ARTIFACT_TYPES } from "@/lib/types/artifact-types";
 import type { Tables } from "@/lib/types/supabase";
 import type { User } from "@supabase/supabase-js";
 import { useState } from "react";
