+++ src/features/missions/components/artifact-form.tsx
@@ -4,11 +4,11 @@ import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Textarea } from "@/components/ui/textarea";
-import { ARTIFACT_TYPES, getArtifactConfig } from "@/lib/artifactTypes";
 import {
   MAX_POSTING_COUNT,
   POSTING_POINTS_PER_UNIT,
 } from "@/lib/constants/mission-config";
+import { ARTIFACT_TYPES, getArtifactConfig } from "@/lib/types/artifact-types";
 import type { Tables } from "@/lib/types/supabase";
 import type { User } from "@supabase/supabase-js";
 import { useState } from "react";
