+++ src/lib/types/artifact-types.ts
